Prefect/docker-requirements.txt
pandas==1.5.2
prefect-gcp[cloud_storage]==0.2.4
pandas-gbq==0.18.1
prefect-sqlalchemy==0.2.2
protobuf==4.21.11
pyarrow==10.0.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.46